Use double instead of int

+++ lib/domain/base_damage.dart
@@ -2,12 +2,13 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:remnant2_calculator/domain/damage_type.dart';
 
 part 'base_damage.freezed.dart';
+
 part 'base_damage.g.dart';
 
 @freezed
 class BaseDamage with _$BaseDamage {
   const factory BaseDamage(
-    int value,
+    double value,
     List<DamageType> damageTypes,
   ) = _BaseDamage;
